+++ src/tools/Built-in/100_other/Keys_and_other/ahkSnippets.data.ts
@@ -218,7 +218,7 @@ export const ahkSnippetsData: string[] = [
     'Ok',
     'Cancel',
     'Abort',
-    'Retry',
+    // 'Retry',
     'Ignore',
     'TryAgain',
     'On',